Add workout history deletion controls
